refactor: Refactor to use MVC pattern
