+++ src/main/java/racing/domain/Round.java
@@ -1,4 +1,4 @@
-package racing.models;
+package racing.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
