+++ src/main/java/racing/domain/Rounds.java
@@ -1,4 +1,4 @@
-package racing.models;
+package racing.domain;
 
 import java.util.ArrayList;
 import java.util.Iterator;
