+++ src/test/java/racing/domain/CarTest.java
@@ -1,4 +1,4 @@
-package racing.models;
+package racing.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
