+++ src/test/java/racing/domain/RoundTest.java
@@ -1,4 +1,4 @@
-package racing.models;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
